First pass at the script
Works on my local machine, untested on the forum server at this point.
It supports enough options to make running it in test mode easy but if
you run it without options it can actually wipe the forum's .htaccess
file so be careful testing it on the live forum server.