Super subtle interpolation bug on the heredoc text
The HTACCESS_TOP and HTACCESS_BOT heredoc sections were interpolating
the contents in the blocks so things like a \ were disappearing from the
final text written to the file and causing the rules to be wrong. It was
super hard to see in the diff with the real .htaccess file on the server
too.

This stops interpolation in those blocks. It also means the problems
were _not_ caused by my mass cleanup of the HTACCESS_TOP block so I
cleaned it back up. Again.