Add error handling
Add some error handling for the responses by requiring a 200 in order to
write the contents of the response body into the htaccess file.

Use the resolv library's IPv4 regex checker against the returned file
contents and only write lines that are valid IPv4 addresses as rules.
Write everything else as a comment line.